gitlint: Lift requirement for Issue-Ref in commit message
Commits are linked by Github to the PR that introduced them.
It is impossible to merge a commit without PR.
Hence, it is sufficient if the PR links to issues.

We have no tooling that would verify that the linkage of a
commit is correct. Thus, making an issue reference in the
commit message mandatory holds no benefit.

It is still allowed to reference an issue in a commit
message. It is simply no longer mandatory.

Issue-ref: see #131
Issue-ref: closes #119